Store the parsed connection parameters
Databases are specified using a DSN string which is parsed into
it's parts and used to establish the database connection.  After
the conneciton is established, the only piece of this that is
saved is the protocol, but for other uses (e.g. opening a database
console) you need to know the connection params.